refactor: merge GitArchive and GitArchiveProcess
I separated the two because I wanted to unit test the GitArchive. It has a test, but that's testing it returns a string. So the value of having a separate git archive object is lost (pretty much). Merge the two. It's being tested in the .feature tests anyway.
